refactor: extract giftPolaroid controller
